chore(oms): move fsku models to domain package
