+++ app/db/base.py
@@ -143,6 +143,7 @@ def init_models(
         "app.tms.records.models",
         "app.tms.billing.models",
         "app.oms.platforms.models",
+        "app.oms.fsku.models",
     ):
         for mod in _iter_model_modules_recursive(pkg_name):
             if mod in ex or mod in loaded:
