+++ app/oms/fsku/models/__init__.py
@@ -1,0 +1,11 @@
+# app/oms/fsku/models/__init__.py
+# Domain-owned ORM models for OMS FSKU.
+
+from app.oms.fsku.models.fsku import Fsku, FskuComponent
+from app.oms.fsku.models.merchant_code_fsku_binding import MerchantCodeFskuBinding
+
+__all__ = [
+    "Fsku",
+    "FskuComponent",
+    "MerchantCodeFskuBinding",
+]
